Move document processors to separate namespace
